// src/shared/enums/SnackbarEnum.ts
export enum SnackbarEnum {
  REQUEST_ERROR = "Ha ocurrido un error. Inténtalo nuevamente",

  REQUEST_SUCCESS = "Se ha guardado la información con éxito.",
  CENTRALIZE_SUCCES = "Se centralizó los 8 Branches con éxito",
  IN_PROGRESS = "Se está ejecutando el proceso",
}

export enum NotificationTypeEnum {
  SUCCESS = "SUCCESS",
  CENTRALIZE_SUCCES = "SUCCESS CEN",
  DESCENTRALIZE_SUCCESS = "SUCCESS DES",
  FAILED = "FAILED",
}